Anomaly RM#109483:PERIOD : anomaly management when we run the close process (#15902)

+++ axelor-account/src/main/java/com/axelor/apps/account/service/period/PeriodServiceAccountImpl.java
@@ -26,7 +26,9 @@ import com.axelor.apps.account.service.move.MoveRemoveService;
 import com.axelor.apps.account.service.move.MoveValidateService;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.db.Period;
+import com.axelor.apps.base.db.TraceBack;
 import com.axelor.apps.base.db.repo.PeriodRepository;
+import com.axelor.apps.base.db.repo.TraceBackRepository;
 import com.axelor.apps.base.db.repo.YearRepository;
 import com.axelor.apps.base.service.AdjustHistoryService;
 import com.axelor.apps.base.service.PeriodServiceImpl;
@@ -35,7 +37,11 @@ import com.axelor.auth.AuthUtils;
 import com.axelor.auth.db.User;
 import com.axelor.db.Query;
 import com.google.inject.Inject;
+import java.util.ArrayList;
+import java.util.List;
 import javax.inject.Singleton;
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.tuple.Pair;
 
 @Singleton
 public class PeriodServiceAccountImpl extends PeriodServiceImpl implements PeriodServiceAccount {
@@ -45,6 +51,10 @@ public class PeriodServiceAccountImpl extends PeriodServiceImpl implements Perio
   protected AccountConfigService accountConfigService;
   protected MoveRemoveService moveRemoveService;
   protected PeriodCheckService periodCheckService;
+  protected TraceBackRepository traceBackRepository;
+
+  protected List<Move> moves = new ArrayList<>();
+  protected int anomalyCount = 0;
 
   @Inject
   public PeriodServiceAccountImpl(
@@ -54,21 +64,29 @@ public class PeriodServiceAccountImpl extends PeriodServiceImpl implements Perio
       MoveRepository moveRepository,
       AccountConfigService accountConfigService,
       MoveRemoveService moveRemoveService,
-      PeriodCheckService periodCheckService) {
+      PeriodCheckService periodCheckService,
+      TraceBackRepository traceBackRepository) {
     super(periodRepo, adjustHistoryService);
     this.moveValidateService = moveValidateService;
     this.moveRepository = moveRepository;
     this.accountConfigService = accountConfigService;
     this.moveRemoveService = moveRemoveService;
     this.periodCheckService = periodCheckService;
+    this.traceBackRepository = traceBackRepository;
   }
 
   @Override
   public void close(Period period) throws AxelorException {
     if (period.getYear().getTypeSelect() == YearRepository.TYPE_FISCAL) {
-      moveValidateService.accountingMultiple(
-          getMoveListByPeriodAndStatusQuery(period, MoveRepository.STATUS_DAYBOOK));
+      Pair<List<Move>, Integer> result =
+          moveValidateService.accountingMultiple(
+              getMoveListByPeriodAndStatusQuery(period, MoveRepository.STATUS_DAYBOOK));
+      moves = result.getLeft();
+      anomalyCount = result.getRight();
       period = periodRepo.find(period.getId());
+    }
+    if (!CollectionUtils.isEmpty(moves)) {
+      return;
     }
     moveRemoveService.deleteMultiple(
         getMoveListByPeriodAndStatusQuery(period, MoveRepository.STATUS_NEW).fetch());
@@ -117,5 +135,33 @@ public class PeriodServiceAccountImpl extends PeriodServiceImpl implements Perio
 
     return super.isClosedPeriod(period)
         && !periodCheckService.isAuthorizedToAccountOnPeriod(period, user);
+  }
+
+  @Override
+  public void closePeriod(Period period) {
+    int oldPeriodStatusSelect = period.getStatusSelect();
+    super.closePeriod(period);
+    if (!CollectionUtils.isEmpty(moves)) {
+      resetStatus(period, oldPeriodStatusSelect);
+    }
+  }
+
+  @Override
+  public List<Move> getMoves() {
+    return moves;
+  }
+
+  @Override
+  public int getAnomalyCount() {
+    return anomalyCount;
+  }
+
+  @Override
+  public List<TraceBack> getAnomalies(String moveIds, int anomalyCount) {
+    return traceBackRepository
+        .all()
+        .filter("self.ref = ?1 AND self.refId IN (" + moveIds + ")", Move.class.getName())
+        .order("-createdOn")
+        .fetch(anomalyCount);
   }
 }
